Implement child process reaping.
This commit implements child process reaping in the event of termination
of a CLI command.

On Windows, the child process is added to a job object that is set to
terminate the child (and its tree) upon the termination of the parent
dotnet process.  On Windows 7 and Server 2008, the dotnet process cannot
already be associated with a job object for the reaping to occur.  On
later Windows versions, a nested job will be created so the reaping will
still occur.  After the child process exits, the job object is closed
without terminating the remaining processes in the job; this allows for
the child process to spawn additional processes that outlive the child.

On POSIX operating systems, a SIGTERM is intercepted and forwarded on to
the child process only. Like the SIGINT forwarding, it is up to the
child process to decide what to do with the SIGTERM signal (the default
is to abort).

Additionally, this fix further expands upon the previous fix to `dotnet
run` to properly handle SIGINT so that all child processes now benefit
from the fixed behavior.  This means MSBuild-forwarding commands like
`dotnet build` now behave as if MSBuild were directly being executed with
respect to Ctrl-C handling.

Fixes #7426.