Import MSBuildUserExtensionsPath\VisualStudioVersion
Historically, per-VS-version customization was possible because imports
under MSBuildUserExtensionsPath were versioned. After moving to
ToolsVersion=Current, that would no longer be possible if we continued
keying off MSBuildToolsVersion.

VisualStudioVersion and MSBuildToolsVersion used to match (from 12.0 to
15.0). So switching to VisualStudioVersion shouldn't be noticeable for
folks that aren't in the minutia of MSBuild--just place the files in the
folder you logically would have anyway.

Closes #4149.