Leave ProjectExtensionsElement if namespace is only diff (#4148)
When setting a ProjectExtensionsElement value, MSBuild diffs the current
and desired XML and writes only in the case that they differ. But this
is producing false negatives because the InnerXml may have a namespace
inherited from the project that isn't represented in the "real" XML.

Allow the current to be equal to the desired _modulo_ that namespace.

Fixes #3904.