Send SIGTERM to process group.
This commit implements sending SIGTERM to the process group if the current
process is the root of a group.  This ensures that all descendant processes
receive the SIGTERM signal.

If the current process is not at the root (e.g. a child of another root), then
only the direct child will be sent the SIGTERM signal.