Fix race in DotnetUnderTest.FullPath on Windows.
There's a race in `DotnetUnderTest.FullPath` on Windows that may result
in the path to dotnet to have a `.exe.exe` suffix when running tests in
parallel.

The root of the problem is that the method is not thread safe and relies
on static state to calculate the result; the state may have already
changed as a result of another thread of execution.

Rather than lock or otherwise change this code, this fix simply removes
the dependency on the static state to do the calculation.  The resulting
method is still not thread safe, but now no longer matters if the full
path is calculated concurrently.