Allow numeric comparison of $(MSBuildToolsVersion) (#4158)
Some extensions (for sure WiX, probably others) compare
$(MSBuildToolsVersion) against a known baseline like "4.0" to determine
further behavior. This is broken by the change of ToolsVersion to
"Current":

```
A numeric comparison was attempted on "$(MSBuildToolsVersion)" that evaluates to "Current" instead of a number, in condition " '$(MSBuildToolsVersion)' == '' OR '$(MSBuildToolsVersion)' < '4.0' ".
```

Allow these comparisons again by treating expansion of that property
specially and as equivalent to CurrentVisualStudioVersion. Those two
values matched for MSBuild 12 to 15, so there shouldn't be too much user
confusion around the change.

Fixes #4150.