Change SIGTERM handling to be more consistent.
This commit addresses code review feedback to the child process reaping change.
It removes the attempt to signal the current process as a process group and
instead makes the behavior consistent regardless of the process group of the
parent dotnet process.

It also makes it more consistent when running the application under `dotnet
run` and running it as an apphost as only the process that is running user code
will receive the SIGTERM; any additionally spawned processes are the
responsibility of the user code to signal.

With these changes, we only SIGTERM the child process.  Additionally, it
introduces a quick wait to see if the process is still present prior to
signaling.