Change command argument in help for `dotnet list reference`.
This commit implements support for the `dotnet list reference` command to
change the name and description of the parent command's argument.

The parent command shows `PROJECT | SOLUTION` for the option name and a
description that suggests solution files are supported.  However, the list
reference command does not accept solutions as arguments.

This takes advantage of the fact that the `Command` class from the command line
parser has a protected method to change the arguments rule of the command; thus
it requires no changes to the command line parser library to implement.