Tolerate unexpandable metadata from definitions in logging (#4186)
Using `ItemDefinitionGroup` metadata definitions, it's possible to create an item with an unexpandable metadatum: the `Include` is not a path-like thing and the metadatum uses a filesystem intrinsic metadatum name like `%(FileName)`.

This causes problems in loggers that try to access `ITaskItem.Value`. This PR teaches them to handle the situation by logging the exception instead of crashing.