Fix flaky SIGINT/SIGTERM tests.
This commit fixes the race condition that caused failures of the SIGINT/SIGTERM
handling tests on non-Windows platforms.

The tests are designed to spawn `dotnet run`, which will itself spawn a child
process.  The tests look for output from the child process; when the child
outputs the needed data, the tests' process will start signaling the dotnet
process.

The race lies between the call to start the process and the attempt to register
the signal handlers in dotnet. If the child process outputs the needed data and
the test process signals the dotnet process *before* the dotnet process has had
a chance to register the signal handlers, then the default signal handlers will
be invoked.  For SIGINT, this results in dotnet exiting with 130 and not
waiting for the child process to terminate.  For SIGTERM, it won't forward the
signal to the child process or exit with the same exit code that the child
exited with.

The fix is to register the handlers prior to the process starting.  On Windows,
we still need to perform an action after the process has started (namely that
the child process is added to the job object) because it needs the handle to
the child process.