Prevent stage0 bundled versions from interfering with the build.
A change to enable the parameterization of the latest known patch numbers for
Microsoft.NETCore.App reused the same property names as those in the generated
bundled versions props file.

When building locally, the stage0 dotnet install was setting the properties
first, resulting in older versions as the fallbacks then what is currently
being used in the repo.

However, when building from an orchestrated build where the properties were
being set globally, those properties get respected and the values are
the expected ones.

This commit fixes this by renaming the properties to those that aren't used in
the bundled versions props file.  The names chosen match those we're using for
the 2.2 builds for consistency.