Fix shutdown race when SIGTERM is received.
This commit fixes a race that results from a SIGTERM signal.

When SIGTERM is received, the signal handling thread will signal the child
process and wait for it to exit.  The main thread will also be waiting on the
child to exit.

When the child exits, there is a race to complete the shutdown of the CLR. This
commit ensures the main thread will block and allow the signal handling thread
to complete the shutdown.