Add a GetPathsOfAllDirectoriesAbove item intrinsic
This splits out part of the functionality of the `GetPathsOfAllItemsAbove` item intrinsic into an intrinsic that simply returns a set of items represents all the directories above the input set of items.

In the context of .editorconfig support in the C# and VB compilers, this is useful as both a stepping stone to computing the set of applicable .editorconfig files, as well as directly providing the set of directories the project systems should watch for the addition or deletion of .editorconfig files.