Prevent copy-local of runtime assets when targeting .NET Framework TFM.
This commit prevents the copying of "runtime" assets (those that end up
in the `runtimes/` sub-directory for resolution based on the RID at
runtime) when building or publishing an application that targets a .NET
Framework TFM.  Only the .NET Core runtime will resolve these assets
based on RID, so copying them for .NET Framework targeted applications
is unnecessary.

Set the `DisableRuntimeTargets` property to `false` to force the copying
of these files.

Fixes dotnet/cli#10979.