Shorter node utilization histogram bars
Limit the number of histogram bar segments. For long runs the number of segments
can be counted in hundreds of thousands (for instance a build which took 8061.7s
would generate a single line line 161,235 characters long) which is a bit
excessive and makes viewing text logs with histograms quite painful at times.

This commit implements a scaled approach in which the number of bar elements
calculated with the original formula is subdivided as follows:

 * <= 100 segments remains as-is, scale 1:1, character used: `.`
 * > 100 <= 1000 segments, scale 1:100, character used: `+`
 * > 1000 segments, scale 1:1000, character used: `#`

Additionaly, scale is printed after the histogram bar if it's different than
`1:1`