Fix publish conflict resolution for runtime pack assets.
This commit fixes a bug in the conflict resolution for runtime pack assets that
was caused by a recent fix to filter out duplicates items in the set of
publish assets introduced by adding the runtime pack assets twice. The
regression was that the adding of the runtime pack assets to the publish assets
was done *after* the conflict resolution occurred; thus, conflicts weren't
detected properly for assets that should conflict with runtime pack assets.

The fix is to always add the runtime pack assets to the list of publish assets
before conflict resolution occurs.  To prevent the duplicate add of the items,
we now filter out the runtime pack assets when adding the
`ReferenceCopyLocalPaths` items to the publish assets.

Fixes #3035.