Fix bash completion when the SDK is not installed
The `dotnet complete` command is an SDK command. If the SDK is not
installed, `dotnet complete` is not available. Triggering completion in
bash without the SDK installed results in errors being shown.

Fix it by checking the exit status of `dotnet complete` before trying to
use to for completions. If the SDK is not installed, we will only show
file completions.

Fixes #11044