Fix web project related tests for 2.2.
A fix was made in 2.2 that included some tests that published a 2.2 web
project and ensured the output was correct for the apphost.  However,
we don't install a 2.2 shared runtime for ASP.NET in our test
environment and therefore one of the tests cannot activate the
framework-dependent apphost.

For the fix, the important thing was to ensure we didn't publish the
fxr/hostpolicy libraries that interfered with the framework-dependent
apphost's activation.  The test still checks for the absence of these
files to cover the original fix.

Additionally, I'm merging the `GivenThatWeWantToPublishAWebProject`
tests into the `GivenThatWeWantToPublishAWebApp`.  The former was
accidentally put into the "Properties" sub-directory anyway.