Devdiv 736904: Fix publishing error when prerequisite location is an absolute location (#4291)
Clickonce publishing fails when a prerequisite component is set to run from a absolute location (for e.g. a network share) instead of the Home Site or Relative to the install folder. This is due to an incorrect check that needs to happen only when the prerequisite component is set to run from the Relative location. This check and the subsequent copy needs to only happen when the prerequisite component is set to run from a location relative to the installation folder.