Use explicit variable for OptProf source branch (#4295)
When the source branch was hardcoded to `$(Build.SourceBranchName)`,
there was no good way to tell the build "build this branch, but take the
latest OptProf information from branch X". That's an important scenario
when standing up a new branch (like a release branch), or trying to
insert updated bits from a feature branch.

This requires a change to the build definition, which is the only place
(currently) that knows what variables to offer at job-scheduling time.