Import publish profiles in the .NET Core SDK.
This commit implements importing publish profiles in the .NET Core SDK.

Previously, there were two ways to get the publish profile respected when
publishing .NET Core applications:

* Using the Web SDK, which implicitly imports the `Microsoft.NET.Sdk.Publish`
SDK.
* Explicitly referencing the `Microsoft.NET.Sdk.Publish` SDK in the `Project
element` (e.g. `Project="Microsoft.NET.Sdk;Microsoft.NET.Sdk.Publish"`).

The problem is that the publish profile was being imported *after* many of the
important SDK properties (such as RuntimeIdentifier, SelfContained, and
UseAppHost) were defaulted by the .NET Core SDK.  Because the .NET Core SDK
does not expect the values of these properties to change once defaulted, users
are unable to set a particular property in the publish profile and get the
expected behavior in the .NET Core SDK.

For example, when setting just the `RuntimeIdentifier` in the publish profile,
it would publish without a RID-subdirectory, not as self-contained, and without
an apphost.  Contrast this to using the `--runtime` option for `dotnet
publish`, which uses a RID-subdirectory and publishes self-contained with an
apphost. Up until now, users had to work around this by setting *all* of the
desired properties in the publish profile that would otherwise be defaulted by
the .NET Core SDK.

The Web SDK fixed this by importing the publish profile *before* the .NET Core
SDK was imported (see aspnet/websdk#534).  However, this could not fix the
problem for users that have an existing project that directly references the
`Microsoft.NET.Sdk.Publish` SDK.

To fix that, the .NET Core SDK will not attempt to import the publish profile
if it hasn't already been imported by the Web SDK.  If the .NET Core SDK
successfully imports the publish profile, it sets a property that prevents the
Web SDK from attempting to import it.

Fixes #10647.