Remove Microsoft.NET.Build.Bundle package
The application bundler was originally intended to be used
as a command line tool via the Microsoft.NET.Build.Bundle package.

However, this plan was subsequently changed.
The bundler is now a library in the Microsoft.NET.HostModel package,
and is directly imported in the SDK repo.

Therefore, remove the Microsoft.NET.Build.Bundle package from the toolset.