Fix ILLink teset flakiness
This change a non-detereministic test failure in the ILLink tests.

There are two changes:

(1)
When creating the test setup for ILLink tests, do not override RestoreAdditionalProjectSources.
Instead add the test reference project to the existing RestoreAdditionalProjectSources.

(2)
The tests create a test-library which is published self-contained.
While the restore is given a rid, it was not marked as self-contained.
The SDK did not restore the platform specific NetCoreAPP likely because
the defaults are differnt for libraries vs executables.