Require VS 2019
Since Arcade pushed to SDK 2.2.203, the easiest thing to do is move to
requiring Visual Studio 2019, which can use that SDK. We'll need to do
so at some point in the future, anyway, to get .NET Core 3.0.

Update toolset dependency to new CLI version, including NuGet, to
support the bootstrap build.