Normalize paths in GetPathsOfAllDirectoriesAbove
The `GetPathsOfAllDirectoriesAbove` item intrinsic takes a set of items and is supposed to return the set of directories above those items--one item per directory. However, it doesn't properly handle some forms of relative paths. For example, if you start with these items:

* C:\alpha\beta\Class1.cs
* C:\alpha\beta\..\Class2.cs

Then you'll get the following output:

* C:\
* C:\alpha
* C:\alpha\beta
* C:\alpha\beta\..

Which means we end up with two different paths that logically represent the same "alpha" directory.

To fix this we first run the item paths through `FileUtilities.NormalizePath(...)` which reduces it to a normalized form without ".." and such.