Don't fail design time builds if asset file doesn't have correct target
This should fix #2322.  The issue is there could be a catch-22 where
there is an assets file with the wrong targets (because the RuntimeIdentifier
or TargetFramework has changed) and we need to do a design-time
build to nominate restore (which will write the right targets to the assets
file), but the build to get the nomination data fails because the assets
file has the wrong targets.