Log error when building for Windows Desktop on non-Windows.
This commit causes the SDK to log an error (NETSDK1100) if the user attempts to
build a Windows Desktop targeted application on a non-Windows platform.

This works via any explicit or implicit (i.e. from the SDK) framework reference
to `Microsoft.WindowsDesktop.App*`; if the current platform is not Windows and
the framework reference exists, the error is logged.

Additionally, this commit prevents the Windows desktop pack from being
downloaded when building a project that doesn't reference the Windows desktop
framework on non-Windows operating systems.

Fixes dotnet/cli#11410 and dotnet/cli#10842.