Fix filtering of satellite assemblies when publishing for netcoreapp3.0.
This commit fixes filtering of satellite assemblies when publishing an
application targeting netcoreapp3.0.

An incorrect property name in the `_FilterSatelliteResourcesForPublish` target
was causing the satellite filtering done for publish to have no effect.

This was only a problem when not reusing the build output for publish;
unfortunately this is currently the case for netcoreapp3.0 targeted
applications due to references that are marked as `PrivateAssets="all"`.

This commit fixes the property name and adds test coverage for publishing
when targeting netcoreapp3.0 with filtered satellite assemblies.

Fixes #3277.