Fix related item output for WinMD native implementation file.
The ResolveAssemblyReference task outputs `RelatedFiles` that should contain
the native implementation file for "native" WinMD references. This is needed to
copy the implementation file locally for P2P references.

However, if the reference to the native WinMD file already has the
`Implementation` metadata set, the task does not output the implementation file
in the `RelatedFiles` output. In the case of C++ projects that output a WinMD
file, the `Implementation` metadata gets set and referencing projects fail to
copy local the implementation file.

This commit changes the behavior so that the task will output the
implementation file even if the `Implementation` metadata is set on the WinMD
reference.

Fixes #4399.