RFC3161 timestamping of clickonce manifest and binaries (#4397)
Currently in ClickOnce project settings, if you enter a RFC 3161 timestamp server URL, timestamping of the manifest and binaries fails. This is because the current code only works with authenthicode timestamp service.

This change will add support for RFC 3161 timestamping for ClickOnce manifests. It will first attempt an RFC 3161 timestamp and it that fails, fall back to using the older authenticode timestamping.

Reviewer: JohnHart;NingLi

Fixes AB#377113.