Explicit opt-in to new resx
Always use the new codepath on .NET Core, but opt into it explicitly on full. This will convert
some cases where a bogus string was getting embedded on Core to a build-time error. I feel ok
about that at the moment, especially since there's the escape hatch to get back to the old
behavior.