Use new ResX reader when finding linked files
This enables incremental up-to-date checks for linked resources on .NET
Core. The implementation is not efficient (it reads the resx file twice,
throwing away all of the information the first time except "are there
ResXFileRefs?"), but that's what the full framework version has done for
years, and the new way should be faster than that.

Fixes #1247.