Artificially increment timestamp in test
The previous test setup failed on Linux on runtimes less than 3.0,
because of https://github.com/dotnet/corefx/issues/32115.

We can't avoid the delay entirely by artificially setting the date into
the future, because we also need the output of the next execution of the
task to have a later date than the previous one, but we can ensure that
the timestamp on the input file is in the future regardless of
granularity.