Fix dotnet store for netcoreapp3.0 (#3345)
Also fix msbuild batching error that caused O(N^2) redundant work, and
flakiness due to race conditions between the redundant workers.

And replace ContinueOnError=WarnAndContinue with just ignoring standard 
error format on exec of crossgen. crossgen prior to 3.0 is notorious for spewing 
noise and the only thing that is reliable is the exit code. The previous approach
around it was way too big of a hammer.