Ignore Ctrl-C for forwarding commands.
This commit ignores Ctrl-C for the remainder of the dotnet process for
forwarding commands and the `run` command which simply spawn the child process
and exit with the child's exit code.

Doing this will prevent a race where SIGINT is handled after the child process
terminates but before the dotnet process terminates.  After the child process
terminates, the handler registered by the process reaper is unregistered.  If a
SIGINT occurs after this, the default signal handler will execute and the
dotnet process will exit with 130 (SIGINT) on POSIX systems.

This should hopefully fix the flakiness observed in the `ItIgnoresSIGINT` test.

Fixes #11088.