Disable flaky test against bug (#3410)
Good news: The test found a product bug (#3409) that it is designed
to catch.

Bad news: It can still pass depending on what's in the nuget cache
from prior tests. So the change that regressed it slipped through
and now we have flaky CI.

Disabling the test against the bug that it caught for now.