Add support for @(InternalsVisibleTo) items that turn into assembly attributes
Given that `InternalsVisibleTo` (IVT) is such a common assembly-level attribute, this adds support
for specifying it directly via simple items, such as:

```
  <ItemGroup>
    <InternalsVisibleTo Include="MyLibrary.Tests" />
  </ItemGroup>
```

Optionally, a `Key` metadata can be specified to provide a strong-named IVT:

```
  <ItemGroup>
    <InternalsVisibleTo Include="MyLibrary.Tests" Key="PUBLIC_KEY" />
  </ItemGroup>
```

The targets will also use automatically a `$(PublicKey)` if available and no `%(Key)` metadata
override is found. Otherwise, it will default to an IVT without a key.

This also avoids having to learn the `_Parameter1` syntax in `AssemblyAttribute` elements, and is
more similar to the way other higher-level properties like `AssemblyTitle` or `Product` are also
turned into assembly attributes.

To turn off this feature, set `$(GenerateInternalsVisibleToAttributes)` to `false`.

Partially fixes #3166