MSBuildGlob Interning Memory Improvement (#4522)
* Interning the culprit & no longer adding trailing slash

The trailing slash was removed because globRoot is exclusively used in conjunction with Path.Combine, which already adds a trailing slash if it doesn't already exist.

* Added the trailing slash

There are many tests that expect a trailing slash be appended to the globRoot. The initial reason for removing the slash was that Path.Combine would add a slash anyway. However, in cases where globRoot combines with an empty string, no trailing slash will be added from Path.Combine. Thus, it stays unless we want a better workaround. This fix provides significant improvement that a more detailed workaround may not even be worth it.