Fix publishing of applications that specify a full path to the profile.
When an application is published in Visual Studio, the full path to the publish
profile is passed as the `WebPublishProfileFile` global property.

Because the .NET Core SDK now imports the publish profile, this caused every
project referenced to also import the publish profile during restore. This may
have unintentional consequences of restoring `netstandard` projects as
`netcoreapp` if a `netcoreapp` TFM is set in the publish profile, for example.

However, during build/publish, the Web SDK is removing `WebPublishProfileFile`
from the global properties for project references.  This causes an error due to
a potential mismatch between the TFM restored and the TFM being
built/published for dependent projects.

The fix is to import the publish profile only when the full path isn't set or
if the project being built is the one being published (i.e. the project to
override project extensions for).

Fixes aspnet/websdk#604.