Add SimpleProjectRootElementCache (#4485)
```ProjectRootElementCache``` is currently the largest source of perf issues in the parsing stage of our internal build engine, holding up anywhere from 35% to 60% of total parse time, almost entirely due to lock contention. It holds a single lock for all operations on the cache as it needs to synchronize access between a dictionary of weak references and LRU cache of strong references, guard against potential concurrent modifications to ```ProjectRootElement``` state, and reload on disk changes. All of these properties are important for the Visual Studio scenario, but not for a pure evaluation where state should not change and the cache is short-lived, and has a large impact on performance when parallelizing project evaluation.

This adds a ```SimpleProjectRootElementCache``` only intended for use in evaluation, which can be enabled via environment variable. It implements the cache through a ConcurrentDictionary and avoids implementing any of the strong/weak reference methods on the original cache. Ideally these would just be removed as a requirement from the base class and  ```SimpleProjectRootElementCache``` would be usable in a full build, but several areas of MSBuild currently seem to take a hard dependency on the strong/weak reference implementation details of the original, so it would take more effort to generalize.