Last installment of Spanifying Expander
Modifications are not pretty, but as many strings as possible were replaced with Spans for added efficiency. More precisely, strings were inefficient because the substring method allocates a new string. The new code, using spans, has a more efficient substring method in refusing to aloocate a new string unless absolutely necessary. If this is deemed a good direction, the next step is to clean up the code.