Fix copying of DiscardBuildResults flag for BuildParameters (#4699)
When I added the `DiscardBuildResults` flag to `BuildParameters`, I forgot to add it to the `Clone` method. So the flag is lost in the `BuildManager` since the first thing it does is clone the parameters. Also added a comment to the effect that we do not serialize this to the nodes since it's not applicable there.