[mono] Add aspnetcore-tooling source from `dotnet/toolset` to ..
.. our `RestoreSourcesOverrides.props`, because this is used for
fetching msbuild SDKs.

The `dotnet/toolset` repo has to do the same thing, but not msbuild.
Since, we bundle SDKs we have this additional requirement and so we are
copying over the newly added source from
https://github.com/dotnet/toolset/blob/4e1936a575cdeddf29fd1662a77216f2ea05800b/NuGet.config#L11 .

The above commit hash is what the updates are being based off of (see
previous commits).