[mono] Workaround for the assumption that `dotnet` is installed (#138)
* [mono] Workaround for the assumption that `dotnet` is installed

We moved to building purely with mono, earlier and thus skip the dotnet
installation in the scripts. This breaks that. Fixing it up in-place for
now.

Commit 2: (Patch on behalf of Alexander Köplinger)
Also, fix the command line, as the `dotnet nuget` uses a different
command line syntax than the standalone `nuget` that mono bundles.

    `nuget locals http-cache -c` -> `nuget locals http-cache -Clear`

* [mono] Workaround: Clear nuget http-cache, if we can

To work around https://github.com/dotnet/arcade/issues/3932 , `nuget
locals http-cache -Clear` is invoked, but in our CI, `nuget` doesn't
seem to be installed in bockbuild's staged prefix.

So, instead we do that only if `dotnet` is available.