[mono] Update Sdk.Web version to match `dotnet/toolset`
We got a newer version from the channel itself
(`3.1.100-preview1.19503.3`), as compared to the one committed to
`dotnet/toolset` `release/3.1.1xx` branch by maestro (`3.0.100-rc2.19465.1`).

The newer version doesn't seem to be available in any of the nuget
sources. Assuming that it becomes available once maestro pushes the
version updates to the repos (`toolset` for instance), updating to what
`toolset` has.