Fix CachingSdkResolverService concurrency.
The documentation for ConcurrentDictionary<T>.GetOrAdd() states:

"If you call GetOrAdd simultaneously on different threads, valueFactory may be called multiple times, but only one key/value pair will be added to the dictionary."

In all of my testing, a particular SDK was never resolved more than once so I didn't think we had a problem.  However, on an NGen'd MSBuild, I noticed that an SDK was being resolved more multiple times.  This is because the GetOrAdd() method does not guarantee that a method is only called once.  Instead, it guarantees that if multiple threads call the method simultaneously they will all receive the same return value.

So the fix is to return a Lazy<T> instead.  This means that mutiple threads could new up multiple instances of Lazy<T> but all threads will receive the same Lazy<T>.  They will all wait for the same result when accessing Value so the SDK will only be resolved once.  The value is then cached for subsequent calls.