Fix tool package install tests
The tests are testing the nuget adapter (giving package name, output location, download a package.)

By the time nuget is from LKG anyway, so stage 0 or stage 2 they are the same. Now stage 2 has the latest nuget, the test should run again stage2.

Swap the implementation of IProjectRestorer, and let it use stage 2