Add TFM check for `Publish-Trimmed-ReadyToRun-SingleFile`
The options `PublishSingleFile`, `PublishReadyToRun`, and `PublishTrimmed` are only supported when targetting netcoreapp3.0 or later.

Trying to publish to a different target (say `.netcoreapp2.1`) today:
* `PublishSingleFile` fails with the `PlaceHolderNotFoundInAppHostException` generated by the `HostWriter`.
* `PublishReadyToRun` and `PublishTrimmed` silently fail to turn on, but the `publish` itself suceeds.

This change adds an explicit TFM check to generate error/warnings for non-conforming targets.

`PublishSingleFile`, requires the following conditions to be true:
* `TargetFramework` is `netcoreapp`
* `TargetFrameworkVersion` is at least `3.0`
* The app is an executable (`OutputType=exe`)
If any of the conditions fail, build fails with an appropriate error

`PublishReadyToRun`, and `PublishTrimmed` require the following conditions to be true:
* `TargetFrameworkVersion` is at least `3.0`
If this condition fails, the build issue warnings, but publish itself succeeds.
The version check should ideally be a failure, but they are warnings for maximum compatibility with `3.0` release.

Fixes #3728