Allow case-varying solution targets (#4822)
Fixes #4811 by switching to the concrete implementation type for the _defaultTargetNames set.

As ISet<T>, the overload resolved for Union was from Linq and did not respect the left-hand
set's comparer. By moving the field to the concrete ImmutableHashSet<T> type, the overload
from that type is used, which does use the collection's comparer (case-insensitive).

When the target names are correctly unioned, the null-reference exception is avoided, because
the dummy target is removed only once, instead of twice (throwing NRE the second time).