[mono] Fix build
Recent upstream merge started copying RuntimeIdentifierGraph.json from
dotnet sdk installed as part of the build. Since, for a mono build we
don't install that, this Copy fails. Instead, we keep a copy of
this and related files in the repo and copy them from there.

Also, update the files from sdk 3.0.100 .