Add Microsoft.Common.props in Current to CoreXT overlay
In the VS repo, some projects attempted to import based on $(MSBuildToolsVersion), and got confused in the '15.0 package with the 16.3 engine' situation.

The right long-term fix is to move to 16.x packages, but for now we can just duplicate this file; all other files seemed to be loaded from the right folders automatically.